Renamed context files to remove the "_context"
